refactor: 컨트롤러와 서비스 사이 Dto 도입

+++ src/main/java/wooteco/subway/line/dao/LineDao.java
@@ -7,6 +7,6 @@ public interface LineDao {
     Line save(Line line);
     Line findOne(Long id);
     List<Line> findAll();
-    void update(int id, Line line);
+    void update(long id, Line line);
     void delete(long id);
 }
